Moving write_derivative_description to bids/layout/utils.py
